refactor: move context to PairSyncService, create SdkClientService
