+++ src/tasks/pair-liquidity-info-history-validator.service.spec.ts
@@ -2,6 +2,7 @@ import { PairLiquidityInfoHistoryValidatorService } from './pair-liquidity-info-
 import { Test, TestingModule } from '@nestjs/testing';
 import { MdwHttpClientService } from '../clients/mdw-http-client.service';
 import { PairLiquidityInfoHistoryDbService } from '../database/pair-liquidity-info-history/pair-liquidity-info-history-db.service';
+import { SdkClientService } from '../clients/sdk-client.service';
 
 const mockMdwClientService = {
   getKeyBlockMicroBlocks: jest.fn(),
@@ -19,6 +20,7 @@ describe('PairLiquidityInfoHistoryValidatorService', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         PairLiquidityInfoHistoryValidatorService,
+        SdkClientService,
         { provide: MdwHttpClientService, useValue: mockMdwClientService },
         {
           provide: PairLiquidityInfoHistoryDbService,
