+++ test/pair-sync.service.spec.ts
@@ -1,59 +1,27 @@
 import { Logger } from '@nestjs/common';
 import { mock } from 'jest-mock-extended';
 
-import { Context } from '../src/lib/contracts';
-import { ContractAddress } from '../src/lib/utils';
 import * as data from './data/context-mockups';
 import * as utils from './utils';
+import { mockContext, mockupEnvVars, TEST_NET_VARS } from './utils';
 import { objSubEv, swapEvent, swapTxInfo } from './data/subscription-events';
 import { PairSyncService } from '../src/tasks/pair-sync.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { PairDbService } from '../src/database/pair/pair-db.service';
 import { TokenDbService } from '../src/database/token/token-db.service';
 import { MdwWsClientService } from '../src/clients/mdw-ws-client.service';
-
-let ctx: ReturnType<typeof utils.mockContext> = null as any;
-
-const initTestContext = (service: PairSyncService) => {
-  type Ev = {
-    onFactory: (ctx: Context) => Promise<void>;
-    refreshPairsLiquidity: (
-      ctx: Context,
-      contract: ContractAddress,
-    ) => Promise<void>;
-    getAllAddresses: () => Promise<ContractAddress[]>;
-  };
-  const { onFactory, refreshPairsLiquidity, getAllAddresses } = mock<Ev>();
-  const logger = mock<Logger>();
-  const eventHandler = service['createOnEventReceived'](
-    ctx,
-    logger,
-    onFactory,
-    refreshPairsLiquidity,
-    getAllAddresses,
-  );
-  return {
-    getAllAddresses,
-    logger,
-    eventHandler,
-    onFactory,
-    refreshPairsLiquidity,
-  };
-};
-type T = ReturnType<typeof initTestContext>;
-let logger: T['logger'] = null as any;
-let eventHandler: T['eventHandler'] = null as any;
-let onFactory: T['onFactory'] = null as any;
-let refreshPairsLiquidity: T['refreshPairsLiquidity'] = null as any;
-let getAllAddresses: T['getAllAddresses'] = null as any;
+import { Context } from '../src/tasks/pair-sync.model';
+import { SdkClientService } from '../src/clients/sdk-client.service';
+import { ContractAddress } from '../src/clients/sdk-client.model';
 
 describe('PairSyncService', () => {
   let service: PairSyncService;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         PairSyncService,
+        SdkClientService,
         { provide: MdwWsClientService, useValue: {} },
         { provide: PairDbService, useValue: {} },
         {
@@ -63,18 +31,52 @@ describe('PairSyncService', () => {
       ],
     }).compile();
     service = module.get<PairSyncService>(PairSyncService);
-
-    ctx = utils.mockContext(data.context2);
-    ({
-      logger,
-      eventHandler,
-      onFactory,
-      refreshPairsLiquidity,
-      getAllAddresses,
-    } = initTestContext(service));
+    await module.init();
   });
 
   describe('createOnEventReceived', () => {
+    let ctx: ReturnType<typeof utils.mockContext> = null as any;
+
+    const initTestContext = (service: PairSyncService) => {
+      type Ev = {
+        onFactory: () => Promise<void>;
+        refreshPairsLiquidity: (contract: ContractAddress) => Promise<void>;
+        getAllAddresses: () => Promise<ContractAddress[]>;
+      };
+      const { onFactory, refreshPairsLiquidity, getAllAddresses } = mock<Ev>();
+      const logger = mock<Logger>();
+      const eventHandler = service['createOnEventReceived'](
+        logger,
+        onFactory,
+        refreshPairsLiquidity,
+        getAllAddresses,
+      );
+      return {
+        getAllAddresses,
+        logger,
+        eventHandler,
+        onFactory,
+        refreshPairsLiquidity,
+      };
+    };
+    type T = ReturnType<typeof initTestContext>;
+    let logger: T['logger'] = null as any;
+    let eventHandler: T['eventHandler'] = null as any;
+    let onFactory: T['onFactory'] = null as any;
+    let refreshPairsLiquidity: T['refreshPairsLiquidity'] = null as any;
+    let getAllAddresses: T['getAllAddresses'] = null as any;
+
+    beforeEach(async () => {
+      ctx = mockContext(data.context2);
+      service.ctx = ctx;
+      ({
+        logger,
+        eventHandler,
+        onFactory,
+        refreshPairsLiquidity,
+        getAllAddresses,
+      } = initTestContext(service));
+    });
     it('ignores unknown types', async () => {
       await eventHandler({
         ...objSubEv,
@@ -141,7 +143,7 @@ describe('PairSyncService', () => {
       await eventHandler({
         ...objSubEv,
       });
-      expect(onFactory).toHaveBeenCalledWith(ctx, 1);
+      expect(onFactory).toHaveBeenCalledWith(1);
     });
 
     it('refresh pair liquidity', async () => {
@@ -180,8 +182,78 @@ describe('PairSyncService', () => {
       await eventHandler({
         ...objSubEv,
       });
-      expect(refreshPairsLiquidity).toHaveBeenCalledWith(ctx, 'ct_p1', 1);
-      expect(refreshPairsLiquidity).toHaveBeenCalledWith(ctx, 'ct_p2', 1);
+      expect(refreshPairsLiquidity).toHaveBeenCalledWith('ct_p1', 1);
+      expect(refreshPairsLiquidity).toHaveBeenCalledWith('ct_p2', 1);
+    });
+  });
+
+  describe('getContext() on testnet', () => {
+    let context: Context | null = null;
+    mockupEnvVars(TEST_NET_VARS);
+
+    beforeAll(async () => {
+      context = await service['getContext']();
+    });
+
+    const ctx = (): Context => {
+      if (!context) {
+        throw 'initiate context first';
+      }
+      return context;
+    };
+
+    it('matches the .env factory', async () => {
+      const { decodedResult: factoryAddress } = await ctx().router.factory();
+      expect(factoryAddress).toBe(process.env.FACTORY_ADDRESS);
+    });
+
+    it('should have at least 16 pairs', async () => {
+      expect(
+        (await ctx().factory.get_all_pairs()).decodedResult.length,
+      ).toBeGreaterThanOrEqual(16);
+    });
+
+    it('pair should return right token addresses', async () => {
+      const { decodedResult: allPairs } = await ctx().factory.get_all_pairs();
+      const pairAddress = allPairs[allPairs.length - 1];
+      expect(pairAddress).toBe(
+        'ct_efYtiwDg4YZxDWE3iLPzvrjb92CJPvzGwriv4ZRuvuTDMNMb9',
+      );
+      const pairMethods = await ctx().getPair(pairAddress);
+      if (!pairMethods) {
+        fail('pairMethods is null');
+      }
+      expect((await pairMethods.token0()).decodedResult).toBe(
+        'ct_7tTzPfvv3Vx8pCEcuk1kmgtn4sFsYCQDzLi1LvFs8T5PJqgsC',
+      );
+      expect((await pairMethods.token1()).decodedResult).toBe(
+        'ct_b7FZHQzBcAW4r43ECWpV3qQJMQJp5BxkZUGNKrqqLyjVRN3SC',
+      );
+    });
+
+    it('regular aex9 token should have right metaInfo', async () => {
+      const tokenMethods = await ctx().getToken(
+        'ct_7tTzPfvv3Vx8pCEcuk1kmgtn4sFsYCQDzLi1LvFs8T5PJqgsC',
+      );
+
+      const { decodedResult: metaInfo } = await tokenMethods.meta_info();
+      expect(metaInfo).toEqual({
+        decimals: 18n,
+        name: 'TestAEX9-B',
+        symbol: 'TAEX9-B',
+      });
+    });
+
+    it('WAE token should have right metaInfo', async () => {
+      const tokenMethods = await ctx().getToken(
+        'ct_JDp175ruWd7mQggeHewSLS1PFXt9AzThCDaFedxon8mF8xTRF',
+      );
+      const { decodedResult: metaInfo } = await tokenMethods.meta_info();
+      expect(metaInfo).toEqual({
+        decimals: 18n,
+        name: 'Wrapped Aeternity',
+        symbol: 'WAE',
+      });
     });
   });
 });
